143-wallet: mobile 不输入号码时金额部分显示默认金额
Ticket: MOSHARE-14285

Root cause: 商品数值使用了float类型，因此默认是.0

Solution: content的类型为int

+++ LetvWallet/src/com/letv/walletbiz/mobile/beans/ProductBean.java
@@ -35,7 +35,7 @@ public class ProductBean implements LetvBaseBean {
         public int product_id;
         public float orig_price;
         public float price;
-        public float content;
+        public int content;
         public String product_name;
         public String sku_sn;
 
@@ -55,7 +55,7 @@ public class ProductBean implements LetvBaseBean {
             return product_name;
         }
 
-        public float getContent(){
+        public int getContent(){
             return content;
         }
     }
